+++ public/index.html
@@ -10,8 +10,6 @@
         <p id="status"></p>
         <button id="reset" class="button">Reset</button>
     </div>
-    <script src="/scripts/requests.js"></script>
-    <script src="/scripts/hangman.js"></script>
-    <script src="/scripts/app.js"></script>
+    <script src="/scripts/bundle.js"></script>
 </body>
 </html>
