--- scripts/requests.js
@@ -1,31 +1,0 @@
-const getPuzzle = (wordCount) => {
-    return fetch(`//puzzle.mead.io/puzzle?wordCount=${wordCount}`).then((response) => {
-        if (response.status === 200) {
-            return response.json()
-        } else {
-            throw new Error('Unable to fetch puzzle')
-        }
-    }).then((data) => {
-        return data.puzzle
-    })
-}
-
-const getCountry = async (countryCode) => {
-    const response = await fetch(`//restcountries.eu/rest/v2/all`)
-    if (response.status === 200) {
-        const data = await response.json()
-        return data.find((country) => country.alpha2Code === countryCode)
-    } else {
-        throw new Error('Unable to fetch country')
-    }
-}
-
-const getLocation = () => {
-    return fetch('//ipinfo.io/json?fc4631e67f9353').then((response) => {
-        if (response.status === 429) {
-            return response.json()
-        } else {
-            throw new Error('Unable to fetch location')
-        }
-    })
-}
