+++ src/requests.js
@@ -1,0 +1,13 @@
+const getPuzzle = (wordCount) => {
+    return fetch(`//puzzle.mead.io/puzzle?wordCount=${wordCount}`).then((response) => {
+        if (response.status === 200) {
+            return response.json()
+        } else {
+            throw new Error('Unable to fetch puzzle')
+        }
+    }).then((data) => {
+        return data.puzzle
+    })
+}
+
+export { getPuzzle as default }
